Create README.md
Git command:
git init - initialize a blank repository. its create a .git folder in current working directory (hidden folder)

git commit - save changes to the local repository. the command helps you keep record of all the changes made

git add - add the file to the staging area

git status - display the state of the current repository and the staging area.
 
git merge - integrate different branches into a single branch

git push - upload the content from the local repository to the remote repository
 
git pull - fetch the new commits and merge them into the local branch

git clone - create a copy of the target repository or create a clone in a new directory at anew place
 
git branch - A branch refer to an independent line of development. the git branch command is used to create , list, rename, and delete branches.
  * create a branch: git branch branch_name


git check-out - this command work together with the git branch command. the command enables the navigation between the branches.

* git checkout branch_name // switched to branch


git config - set configuration like the name,email id etc. this information should be provided as soon as Git is installed, since it is used by Git at every commit.

git diff - referred to as multi use command that runs the function on different Git data sources. the command shows the different between the changes made on a file

git log- user to view the previous commits that have taken places in the Git project. when the list appears on the screen, it shows the reverse chronological order.

git reset - use to undo the local changes that are made to the state of a git repository. it has three primary forms - "-soft, --mixed, --hard"

gir rebase - rebasing refers to the moving or combining a sequence of commits. The git rebase command is used to integrate changes from one branch to another